feat(i18n): Expand error catalog from 80 to 164 codes with full i18n support
Add 84 new error codes across all compiler phases with 100% Korean translation coverage:

Phase 1: Design & Testing
- Extended i18n/__init__.spl with 84 English error constants
- Extended i18n/__init__.ko.spl with 84 Korean translations  
- Created 95 comprehensive SSpec test files
- Fixed src/i18n/Cargo.toml to enable simple-format by default

Phase 2: Infrastructure Implementation
- Added 70+ new error code constants to src/compiler/src/error.rs
- Added 68 new error entries to i18n/compiler.spl
- Added 68 new i18n conversion cases to src/compiler/src/i18n_diagnostics.rs
- Implemented helper extraction functions for error context

New Error Codes:
- E1011-E1080: Extended semantic errors (70 codes)
- E1104-E1105: Control flow errors (2 codes)
- E1301-E1302: Capability errors (2 codes)  
- E1403: Macro keyword error (1 code)
- E2003-E2009: Codegen errors (7 codes)
- E3006-E3009: Runtime errors (4 codes)
- E4005: FFI error (1 code)

Testing:
- All 21 i18n tests passing
- Compiler compiles successfully
- 95 SSpec tests created (113% coverage target)

Total catalog now contains 164 error codes ready for compiler integration.

+++ src/compiler/src/codegen/instr/mod.rs
@@ -183,7 +183,7 @@ pub fn compile_instruction<M: Module>(
         }
 
         MirInst::InterpEval { dest, expr_index } => {
-            compile_interp_eval(ctx, builder, *dest, *expr_index)?;
+            compile_interp_eval(ctx, builder, *dest, *expr_index as usize)?;
         }
 
         MirInst::ArrayLit { dest, elements } => {
@@ -391,7 +391,7 @@ pub fn compile_instruction<M: Module>(
             byte_offset,
             field_type,
         } => {
-            compile_field_get(ctx, builder, *dest, *object, *byte_offset, *field_type)?;
+            compile_field_get(ctx, builder, *dest, *object, *byte_offset as usize, *field_type)?;
         }
 
         MirInst::FieldSet {
@@ -400,7 +400,7 @@ pub fn compile_instruction<M: Module>(
             field_type,
             value,
         } => {
-            compile_field_set(ctx, builder, *object, *byte_offset, *field_type, *value)?;
+            compile_field_set(ctx, builder, *object, *byte_offset as usize, *field_type, *value)?;
         }
 
         MirInst::MethodCallStatic {
